Updated output in office repair

+++ src/simModel/Output.java
@@ -18,21 +18,21 @@ public class Output {
 
 	protected double getSatisfactionLevelT12() {
 		if (totalNumberT12Contracts != 0) {
-			return contractsT12satisfied / totalNumberT12Contracts;
+			return (float)contractsT12satisfied / (float)totalNumberT12Contracts;
 		}
 		return 0;
 	}
 
 	protected double getSatisfactionLevelT34() {
 		if (totalNumberT34Contracts != 0) {
-			return contractsT34satisfied / totalNumberT34Contracts;
+			return (float)contractsT34satisfied / (float)totalNumberT34Contracts;
 		}
 		return 0;
 	}
 
 	protected double getSatisfactionLevelAll() {
 		if (totalNumberT12Contracts + totalNumberT34Contracts != 0) {
-			return (contractsT12satisfied + contractsT34satisfied) / (totalNumberT12Contracts + totalNumberT34Contracts);
+			return (float)(contractsT12satisfied + contractsT34satisfied) / (float)(totalNumberT12Contracts + totalNumberT34Contracts);
 		}
 		return 0;
 	}
